UX-315 Move algolia query to its own file
