+++ site/gatsby-config.js
@@ -1,44 +1,5 @@
 require('dotenv').config();
 
-function pageToAlgoliaRecord({ node: { id, frontmatter, fields, ...rest } }) {
-  return {
-    objectID: id,
-    ...frontmatter,
-    ...fields,
-    ...rest
-  };
-}
-
-const pageQuery = `{
-  allMdx {
-    edges {
-      node {
-        id
-        fields {
-          slug
-        }
-        frontmatter {
-          title
-          category
-        }
-        excerpt(pruneLength: 200)
-      }
-    }
-  }
-}`;
-
-const queries = [
-  {
-    query: pageQuery,
-    transformer: ({ data }) => {
-      return data.allMdx.edges.map(pageToAlgoliaRecord);
-    }, // optional
-    indexName: 'matchbox', // overrides main index name, optional
-    settings: { attributesToSnippet: [`excerpt:20`] }
-    // matchFields: ['slug', 'modified'] // Array<String> overrides main match fields, optional
-  }
-];
-
 module.exports = {
   siteMetadata: {
     title: 'Matchbox',
@@ -143,7 +104,7 @@ module.exports = {
       options: {
         appId: process.env.GATSBY_ALGOLIA_APP_ID,
         apiKey: process.env.ALGOLIA_ADMIN_KEY,
-        queries: queries
+        queries: require('./src/utils/algolia-queries')
       }
     }
     // this (optional) plugin enables Progressive Web App + Offline functionality
