--- site/src/components/utils/algolia-queries.js
@@ -1,65 +1,0 @@
-// function pageToAlgoliaRecord({ node: { id, frontmatter, fields, ...rest } }) {
-//   return {
-//     objectID: id,
-//     ...frontmatter,
-//     ...fields,
-//     ...rest
-//   };
-// }
-
-// const pageQuery = `{
-//   allMdx {
-//     edges {
-//       node {
-//         id
-//         fields {
-//           slug
-//         }
-//         frontmatter {
-//           title
-//         }
-//         excerpt(pruneLength: 200)
-//       }
-//     }
-//   }
-// }`;
-
-// const queries = [
-//   {
-//     query: pageQuery,
-//     transformer: ({ data }) => {
-//       return data.allMdx.edges.map(pageToAlgoliaRecord);
-//     }, // optional
-//     indexName: 'matchbox', // overrides main index name, optional
-//     settings: { attributesToSnippet: [`excerpt:20`] }
-//     // matchFields: ['slug', 'modified'] // Array<String> overrides main match fields, optional
-//   }
-// ];
-
-// module.exports = queries;
-
-const myQuery = `{
-  allSitePage {
-    edges {
-      node {
-        objectID: id
-        component
-        path
-      }
-    }
-  }
-}`;
-
-const queries = [
-  {
-    query: myQuery,
-    transformer: ({ data }) => data.allSitePage.edges.map(({ node }) => node), // optional
-    indexName: 'index name to target', // overrides main index name, optional
-    settings: {
-      // optional, any index settings
-    },
-    matchFields: ['slug', 'modified'] // Array<String> overrides main match fields, optional
-  }
-];
-
-module.exports = queries;
